refactor: decouple resume repository
