+++ src/libs/modules/database/repositories/resume/resume.interface.ts
@@ -1,0 +1,9 @@
+export interface ResumeRepository {
+  findFirst;
+  delete;
+  update;
+  findMany;
+  count;
+}
+
+export const ResumeRepository = Symbol('ResumeRepository');
